LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/color.tentative-expected.txt


PASS Empty value should return #000000
PASS Missing value should return #000000
PASS Valid simple color: should return #ffffff
PASS Valid simple color (containing LATIN CAPITAL LETTERS): should return #ffffff (converted to ASCII lowercase)
PASS Zero-padding
FAIL Invalid simple color: not 7 characters long assert_equals: expected "#000000" but got "#ffffff"
PASS Invalid simple color: no starting # sign
PASS Invalid simple color: non ASCII hex digits
PASS Invalid simple color: foobar
PASS Invalid color: trailing Null (U+0000)
PASS Invalid color: trailing ;
FAIL Invalid color: leading space assert_equals: expected "#000000" but got "#ffffff"
FAIL Invalid color: trailing space assert_equals: expected "#000000" but got "#ffffff"
FAIL Invalid color: leading+trailing spaces assert_equals: expected "#000000" but got "#ffffff"
FAIL Invalid color: keyword crimson assert_equals: expected "#000000" but got "#dc143c"
FAIL Invalid color: keyword bisque assert_equals: expected "#000000" but got "#ffe4c4"
PASS Invalid color: keyword currentColor
PASS Invalid color: keyword transparent
FAIL Invalid color: keyword ActiveBorder assert_equals: expected "#000000" but got "#ffffff"
PASS Invalid color: keyword inherit
FAIL Invalid color: rgb(1,1,1) assert_equals: expected "#000000" but got "#010101"
FAIL Invalid color: rgb(1,1,1,1) assert_equals: expected "#000000" but got "#010101"
PASS Invalid color: PILE OF POO (U+1F4A9)
